refactor: 레이어별 패키지로 파일 이동
domain 이동
- factory로 팩토리 이동
- entity로 모델 이동
- generator로  생성기 이동

ui 이동
- output
- input
- controller 이동
명확한 책임 분리보다는 객체한테 어울리는 기능을 가지고 있는가로 분류해서 추후 수정할 예정
