+++ src/main/kotlin/racingcar/domain/generator/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain.generator
 
 interface NumberGenerator {
     fun generate(): Int
