+++ src/main/kotlin/racingcar/ui/UserInput.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.ui
 
 import camp.nextstep.edu.missionutils.Console
