+++ src/test/kotlin/racingcar/CarFactoryTest.kt
@@ -3,6 +3,8 @@ package racingcar
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
+import racingcar.domain.entity.Car
+import racingcar.domain.factory.CarFactory
 
 class CarFactoryTest {
