+++ src/test/kotlin/racingcar/CarTest.kt
@@ -2,6 +2,7 @@ package racingcar
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
+import racingcar.domain.entity.Car
 
 class CarTest {
